Created an abstract Installer class and moved core functions over
